organized types (non-related)
