+++ src/components/features/AI-Planner/StudyPlan.tsx
