+++ src/types/ai-planner/debugPlan.ts
