+++ src/types/ai-planner/studyDay.ts
