+++ src/types/ai-planner/validPlan.ts
