--- src/types/items.ts
@@ -1,6 +1,0 @@
-export interface Items {
-    url: string;
-    name: string;
-    category: string;
-    worth: number
-}
