--- src/types/schoolitem.ts
@@ -1,8 +1,0 @@
-export interface SchoolItemProps {
-    _id: string;
-    name: string;
-    location: string;
-    desc: string;
-    joined: string;
-    img: string;
-}
