--- src/types/wrapper.ts
@@ -1,3 +1,0 @@
-export interface Wrapper {
-    children: React.ReactNode; 
-}
